delete main.py and replace by server.py : no thread allowed in Android
