+++ server/main.py
@@ -1,43 +1,108 @@
-import sys
-
-from threading import Thread, Lock
-import logging
+import os
+from flask import Flask, url_for, render_template, jsonify, request, make_response
 import webview
-from time import sleep
-from server import run_server
+import webbrowser
+import app
 
-server_lock = Lock()
+pages_dir = os.path.join(os.getcwd(), "pages")  # development path
+if not os.path.exists(pages_dir):  # frozen executable path
+    pages_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), "pages")
 
-logger = logging.getLogger(__name__)
+server = Flask(__name__, static_folder=pages_dir, template_folder=pages_dir)
+server.config["SEND_FILE_MAX_AGE_DEFAULT"] = 1  # disable caching
 
 
-def url_ok(url, port):
-    # Use httplib on Python 2
-    try:
-        from http.client import HTTPConnection
-    except ImportError:
-        from httplib import HTTPConnection
+@server.after_request
+def add_header(response):
+    response.headers['Cache-Control'] = 'no-store'
+    return response
 
-    try:
-        conn = HTTPConnection(url, port)
-        conn.request("GET", "/")
-        r = conn.getresponse()
-        return r.status == 200
-    except:
-        logger.exception("Server not started")
-        return False
 
-def start_server():
-    logger.debug("Starting server")
-    t = Thread(target=run_server)
-    t.daemon = True
-    t.start()
-    logger.debug("Checking server")
+@server.route("/init")
+def initialize():
+    """
+    Perform heavy-lifting initialization asynchronously.
+    :return:
+    """
+    can_start = app.initialize()
 
-    while not url_ok("127.0.0.1", 23948):
-        sleep(0.1)
+    if can_start:
+        response = {
+            "status": "ok",
+        }
+    else:
+        response = {
+            "status": "error"
+        }
 
-    logger.debug("Server started")
+    return jsonify(response)
 
-if __name__ == '__main__':
-    start_server()
+
+@server.route("/")
+def landing():
+    """
+    Render index.html. Initialization is performed asynchronously in initialize() function
+    """
+    return render_template("index.html")
+
+
+@server.route('/', defaults={'path': ''})
+@server.route('/<path:path>')
+def other_pages(path):
+    """
+    Catch all other URLs that are not already caught by specific cases, this enables link clicking.
+    """
+    return render_template(path)
+
+
+@server.route("/choose/path")
+def choose_path():
+    """
+    Invoke a folder selection dialog here
+    :return:
+    """
+    dirs = webview.create_file_dialog(webview.FOLDER_DIALOG)
+    if dirs and len(dirs) > 0:
+        directory = dirs[0]
+        if isinstance(directory, bytes):
+            directory = directory.decode("utf-8")
+
+        response = {"status": "ok", "directory": directory}
+    else:
+        response = {"status": "cancel"}
+
+    return jsonify(response)
+
+
+@server.route("/fullscreen")
+def fullscreen():
+    webview.toggle_fullscreen()
+    return jsonify({})
+
+
+@server.route("/open-url", methods=["POST"])
+def open_url():
+    url = request.json["url"]
+    webbrowser.open_new_tab(url)
+
+    return jsonify({})
+
+
+@server.route("/do/stuff")
+def do_stuff():
+    result = app.do_stuff()
+
+    if result:
+        response = {"status": "ok", "result": result}
+    else:
+        response = {"status": "error"}
+
+    return jsonify(response)
+
+
+def run_server():
+    server.run(host="127.0.0.1", port=23948, threaded=True)
+
+
+if __name__ == "__main__":
+    run_server()
